mcp-python/pyproject.toml
[project]
name = "mcp-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "pygithub>=2.6.1",
]
